[#22] Controller Request / Response 객체 수정
*Request, *Response 단일 객체 내 요청 별 내부 클래스 사용
