+++ src/main/java/kr/hhplus/be/server/interfaces/api/coupon/CouponApiController.java
@@ -9,16 +9,16 @@ public class CouponApiController implements CouponApiSpec{
 
     @Override
     @PostMapping
-    public ResponseEntity<CouponCreateResponse> createCoupon(
-            @RequestBody CouponCreateRequest request) {
-        return ResponseEntity.ok(new CouponCreateResponse());
+    public ResponseEntity<CouponResponse.Create> createCoupon(
+            @RequestBody CouponRequest.Create request) {
+        return ResponseEntity.ok(new CouponResponse.Create());
     }
 
     @Override
     @PostMapping("/{id}/issue")
-    public ResponseEntity<CouponIssueResponse> issueCoupon(
+    public ResponseEntity<CouponResponse.Issue> issueCoupon(
             @PathVariable("id") Long couponId,
             @RequestBody Long userId) {
-        return ResponseEntity.ok(new CouponIssueResponse());
+        return ResponseEntity.ok(new CouponResponse.Issue());
     }
 }
