+++ src/main/java/kr/hhplus/be/server/interfaces/api/coupon/CouponApiSpec.java
@@ -16,17 +16,17 @@ public interface CouponApiSpec {
     @Operation(summary = "쿠폰 생성", description = "새로운 쿠폰을 생성합니다.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "쿠폰 생성 성공",
-            content = {@Content(mediaType = "application/json", schema = @Schema(implementation = CouponCreateResponse.class))})
+            content = {@Content(mediaType = "application/json", schema = @Schema(implementation = CouponResponse.Create.class))})
     })
-    ResponseEntity<CouponCreateResponse> createCoupon(CouponCreateRequest request);
+    ResponseEntity<CouponResponse.Create> createCoupon(CouponRequest.Create request);
 
 
     @Operation(summary = "사용자 쿠폰 발급", description = "사용자 쿠폰을 발급합니다.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "쿠폰 생성 성공",
-                    content = {@Content(mediaType = "application/json", schema = @Schema(implementation = CouponIssueResponse.class))})
+                    content = {@Content(mediaType = "application/json", schema = @Schema(implementation = CouponResponse.Issue.class))})
     })
-    ResponseEntity<CouponIssueResponse> issueCoupon(
+    ResponseEntity<CouponResponse.Issue> issueCoupon(
         Long couponId,
         @RequestBody(content =
         @Content(mediaType = "application/json", examples = {@ExampleObject(name = "사용자 고유번호", value = "10")})) Long userId);
