+++ src/main/java/kr/hhplus/be/server/interfaces/api/item/ItemResponse.java
@@ -4,47 +4,117 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public class ItemResponse {
 
-    @Schema(name = "itemId", description = "아이템 고유 번호", example = "1")
-    private Long itemId;
+    static public class Detail{
 
-    @Schema(name = "name", description = "상품 명", example = "자전거")
-    private String name;
+        protected Detail(){}
 
-    @Schema(name = "price", description = "상품 가격", example = "10000")
-    private Integer price;
+        @Schema(name = "itemId", description = "아이템 고유 번호", example = "1")
+        private Long itemId;
 
-    @Schema(name = "stock", description = "상품 재고", example = "5")
-    private Integer stock;
+        @Schema(name = "name", description = "상품 명", example = "자전거")
+        private String name;
 
-    public Long getItemId() {
-        return itemId;
+        @Schema(name = "price", description = "상품 가격", example = "10000")
+        private Integer price;
+
+        @Schema(name = "stock", description = "상품 재고", example = "5")
+        private Integer stock;
+
+        public Detail(Long itemId, String name, Integer price, Integer stock) {
+            this.itemId = itemId;
+            this.name = name;
+            this.price = price;
+            this.stock = stock;
+        }
+
+        public Long getItemId() {
+            return itemId;
+        }
+
+        public void setItemId(Long itemId) {
+            this.itemId = itemId;
+        }
+
+        public String getName() {
+            return name;
+        }
+
+        public void setName(String name) {
+            this.name = name;
+        }
+
+        public Integer getPrice() {
+            return price;
+        }
+
+        public void setPrice(Integer price) {
+            this.price = price;
+        }
+
+        public Integer getStock() {
+            return stock;
+        }
+
+        public void setStock(Integer stock) {
+            this.stock = stock;
+        }
+
     }
 
-    public void setItemId(Long itemId) {
-        this.itemId = itemId;
-    }
+    static public class Rank{
 
-    public String getName() {
-        return name;
-    }
+        protected Rank() {
+        }
 
-    public void setName(String name) {
-        this.name = name;
-    }
+        @Schema(name = "itemId", description = "상품 고유번호", example = "10")
+        private Long itemId;
 
-    public Integer getPrice() {
-        return price;
-    }
+        @Schema(name = "name", description = "상품명", example = "자전거")
+        private String name;
 
-    public void setPrice(Integer price) {
-        this.price = price;
-    }
+        @Schema(name = "price", description = "상품 가격", example = "10000")
+        private Integer price;
 
-    public Integer getStock() {
-        return stock;
-    }
+        @Schema(name = "orderCount", description = "3일간 총 판매 수", example = "55213")
+        private Integer orderCount;
 
-    public void setStock(Integer stock) {
-        this.stock = stock;
+        public Rank(Long itemId, String name, Integer price, Integer orderCount) {
+            this.itemId = itemId;
+            this.name = name;
+            this.price = price;
+            this.orderCount = orderCount;
+        }
+
+        public Long getItemId() {
+            return itemId;
+        }
+
+        public void setItemId(Long itemId) {
+            this.itemId = itemId;
+        }
+
+        public String getName() {
+            return name;
+        }
+
+        public void setName(String name) {
+            this.name = name;
+        }
+
+        public Integer getPrice() {
+            return price;
+        }
+
+        public void setPrice(Integer price) {
+            this.price = price;
+        }
+
+        public Integer getOrderCount() {
+            return orderCount;
+        }
+
+        public void setOrderCount(Integer orderCount) {
+            this.orderCount = orderCount;
+        }
     }
 }
