--- src/main/java/kr/hhplus/be/server/interfaces/api/user/UserCouponResponseWrapper.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.user;
-
-import java.util.List;
-
-public class UserCouponResponseWrapper {
-
-    List<UserCouponResponse> userCouponList;
-}
