+++ src/test/java/kr/hhplus/be/server/controller/OrderApiControllerTest.java
@@ -1,8 +1,6 @@
 package kr.hhplus.be.server.controller;
 
-import kr.hhplus.be.server.interfaces.api.order.OrderApiController;
-import kr.hhplus.be.server.interfaces.api.order.OrderCreateRequest;
-import kr.hhplus.be.server.interfaces.api.order.OrderCreateResponse;
+import kr.hhplus.be.server.interfaces.api.order.*;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -18,19 +16,19 @@ public class OrderApiControllerTest {
     @Autowired
     OrderApiController orderApiController;
 
-    OrderCreateResponse orderCreateResponse;
-    OrderCreateRequest orderCreateRequest;
+    OrderResponse.Create orderCreateResponse;
+    OrderRequest.Create orderCreateRequest;
 
     @BeforeEach
     void setUp() {
-        orderCreateResponse = new OrderCreateResponse();
-        orderCreateRequest = new OrderCreateRequest();
+        orderCreateResponse = new OrderResponse.Create();
+        orderCreateRequest = new OrderRequest.Create();
     }
 
     @DisplayName("주문 생성 성공 (HTTP STATUS 200)")
     @Test
     void test(){
-        ResponseEntity<OrderCreateResponse> order = orderApiController.createOrder(orderCreateRequest);
+        ResponseEntity<OrderResponse.Create> order = orderApiController.createOrder(orderCreateRequest);
         Assertions.assertEquals(HttpStatus.OK, order.getStatusCode());
     }
