+++ src/test/java/kr/hhplus/be/server/controller/UserApiControllerTest.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.controller;
 
 import kr.hhplus.be.server.interfaces.api.user.UserApiController;
-import kr.hhplus.be.server.interfaces.api.user.UserCouponResponse;
 import kr.hhplus.be.server.interfaces.api.user.UserResponse;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -24,28 +23,28 @@ public class UserApiControllerTest {
     @DisplayName("사용자 조회 성공(HTTP STATUS 200)")
     @Test
     void test() {
-        ResponseEntity<UserResponse> byId = userApiController.findById(1L);
+        ResponseEntity<UserResponse.Detail> byId = userApiController.findById(1L);
         Assertions.assertEquals(HttpStatus.OK, byId.getStatusCode());
     }
 
     @DisplayName("사용자 포인트 조회 성공(HTTP STATUS 200)")
     @Test
     void test2(){
-        ResponseEntity<UserPointResponse> userPointById = userApiController.findUserPointById(1L);
+        ResponseEntity<UserResponse.UserPoint> userPointById = userApiController.findUserPointById(1L);
         Assertions.assertEquals(HttpStatus.OK, userPointById.getStatusCode());
     }
 
     @DisplayName("사용자 쿠폰 조회 성공(HTTp STATUS 200)")
     @Test
     void test3(){
-        ResponseEntity<UserCouponResponse> userCouponListByUserId = userApiController.findUserCouponListByUserId(1L);
+        ResponseEntity<UserResponse.UserCoupon> userCouponListByUserId = userApiController.findUserCouponListByUserId(1L);
         Assertions.assertEquals(HttpStatus.OK, userCouponListByUserId.getStatusCode());
     }
 
     @DisplayName("사용자 포인트 충전 성공(HTTP STATUS 200)")
     @Test
     void test4(){
-        ResponseEntity<UserPointResponse> userPointResponseResponseEntity = userApiController.chargePoint(1L, 2000);
+        ResponseEntity<UserResponse.UserPoint> userPointResponseResponseEntity = userApiController.chargePoint(1L, 2000);
         Assertions.assertEquals(HttpStatus.OK, userPointResponseResponseEntity.getStatusCode());
     }
